Migrate from Motorola FFP to IEEE 754 single-precision floats
Complete migration of ACE's floating-point system from Motorola Fast
Floating Point (FFP) to IEEE 754 single-precision format. This enables
native C float types in the runtime, correct float handling on all
Amiga hardware with OS 2.0+, and eliminates the opaque-long workaround.

Key changes:
- Compiler: use mathieeesingbas/singtrans libraries, IEEESPxxx functions
- Startup: open IEEE libraries instead of FFP
- Runtime C: rewrite floatstr.c (was ffpstr.c), update val.c, pow.c
- Runtime ASM: update all LVO refs and IEEE constants in fmath.s (was
  ffp.s), gfx.s, misc.s, time.s, scrwin.s, sound.s, turtle.s
- Rename all internal FFP identifiers: gen_ffp_call->gen_float_call,
  mathffpused->mathfloatused, _openmathffp->_openmathieee, etc.
- dp-float: remove FFP<->IEEE bridge (direct IEEE now)
- Docs: update ref.txt, ace.txt, quickref.txt for IEEE format

+++ src/ace/c/lexvar.c
@@ -157,7 +157,7 @@ char   ut_id[MAXIDSIZE];	 /* untouched identifier for DATA statement */
 char   stringval[MAXSTRLEN];     /* string literal */
 SHORT  shortval;                 /* short literal */
 LONG   longval;                  /* long literal */
-LONG  singleval;                /* single precision literal */
+float singleval;                /* single precision literal */
 double doubleval;                /* double precision literal */ 
 SHORT  dimsize[255];		 /* array indices holder */
